Use brew for mac

+++ install.sh
@@ -91,10 +91,13 @@ start() {
       done
       
       # finalize
+      pushd $HOME/.c9/node_modules/.bin
+
       for FILE in $HOME/.c9/node_modules/.bin/* 
       do
-          perl -i -p -e 's/#!\/usr\/bin\/env node/#!'${NODE//\//\\\/}'/' $(readlink -f $FILE)
+          perl -i -p -e 's/#!\/usr\/bin\/env node/#!'${NODE//\//\\\/}'/' $(readlink $FILE)
       done
+      popd
       
       echo 1 > $HOME/.c9/installed
       echo :Done.
@@ -130,13 +133,17 @@ node(){
 
 tmux(){
   echo :Installing TMUX
+  if [ $os = "darwin" ]; then
+    brew install tmux
+    ln -s $(which tmux) ~/.c9/bin/tmux
+  else
+    curl -sSOL https://raw.github.com/c9/install/fix-tar/packages/tmux/tmux-$1-$2.tar.gz
+    tar xvfz tmux-$1-$2.tar.gz
+    rm tmux-$1-$2.tar.gz
 
-  curl -sSOL https://raw.github.com/c9/install/fix-tar/packages/tmux/tmux-$1-$2.tar.gz
-  tar xvfz tmux-$1-$2.tar.gz
-  rm tmux-$1-$2.tar.gz
-
-  rm -f ~/.c9/bin/tmux
-  ln -s ~/.c9/local/bin/tmux ~/.c9/bin/tmux
+    rm -f ~/.c9/bin/tmux
+    ln -s ~/.c9/local/bin/tmux ~/.c9/bin/tmux
+  fi
 }
 
 nak(){
